refactor: extract activist query into shared module
