+++ pkg/activists/columns_meta.go
@@ -1,4 +1,4 @@
-package model
+package activists
 
 import (
 	"database/sql"
