+++ pkg/shared/events.go
@@ -1,0 +1,5 @@
+package shared
+
+// ActionEventTypesSQL is the SQL-quoted list of event types that count as
+// "actions" for activist metrics (last action, MPP requirements, etc.).
+const ActionEventTypesSQL string = "'Outreach', 'Action', 'Campaign Action', 'Animal Care', 'Frontline Surveillance'"
